no libraries per project
unless we're saying a programming language or framework is just an
application, then applications may have libraries. but in that case
there's no distinction between libraries and plugins.

in fact, we may model programming language modules as plugins, because
they probably have the same properties, such as method of composition