usage/fix(sylk): mount *from* external dirs, not the opposite
A service should not bind mount `StateDirectory=`s
onto external dirs, because:

1. When the service is stopped those external dirs no longer contain what's expected.
Whereas they will contain the same if the opposite is done.

2. If the service is ever hardened with `PrivateMounts=`
bind-mounted directories will not be reachable from the host without
entering the mount namespace and `RootDirectory=` of the service.
Whereas nothing will change for users if the opposite is done.

The pattern followed is:

```nix
{
  options = {
    systemd.services.${service}.settings.config.Server.spool_dir = lib.mkOption {
       default = "/var/lib/${service}";
    };
  };
  config = {
    systemd.services.${service}.serviceConfig = {
       BindPaths = [ cfg.settings.config.Server.spool_dir ];
       StateDirectory = [ service ];
    };
  };
}
```
Note that it still uses `StateDirectory=` to create the default,
even if not used when the user changes it.