maint/clarify(sylk): factorize service's name
It's not enough to allow multiple instances of the service,
but it's clearer IMHO, especially when we context-switch
between services, kinda like `cfg` instead of `config.services.foo`.

Note that `LogsDirectory=` does not need to be a list here,
NixOS will transform it into a list if need be when merging options.